+++ bmc-database/src/main/java/com/oracle/bmc/database/DatabaseAsync.java
@@ -1348,7 +1348,9 @@ public interface DatabaseAsync extends AutoCloseable {
                     handler);
 
     /**
-     * Gets a list of supported Autonomous Database versions. Note that preview version software is only available for [serverless deployments](https://docs.cloud.oracle.com/Content/Database/Concepts/adboverview.htm#AEI).
+     * Gets a list of supported Autonomous Database versions. Note that preview version software is only available for
+     * databases with [shared Exadata infrastructure](https://docs.cloud.oracle.com/Content/Database/Concepts/adboverview.htm#AEI).
+     *
      *
      * @param request The request object containing the details to send
      * @param handler The request handler to invoke upon completion, may be null.
@@ -1364,6 +1366,22 @@ public interface DatabaseAsync extends AutoCloseable {
                                     ListAutonomousDbPreviewVersionsRequest,
                                     ListAutonomousDbPreviewVersionsResponse>
                             handler);
+
+    /**
+     * Gets a list of supported Autonomous Database versions.
+     *
+     * @param request The request object containing the details to send
+     * @param handler The request handler to invoke upon completion, may be null.
+     * @return A Future that can be used to get the response if no AsyncHandler was
+     *         provided. Note, if you provide an AsyncHandler and use the Future, some
+     *         types of responses (like java.io.InputStream) may not be able to be read in
+     *         both places as the underlying stream may only be consumed once.
+     */
+    java.util.concurrent.Future<ListAutonomousDbVersionsResponse> listAutonomousDbVersions(
+            ListAutonomousDbVersionsRequest request,
+            com.oracle.bmc.responses.AsyncHandler<
+                            ListAutonomousDbVersionsRequest, ListAutonomousDbVersionsResponse>
+                    handler);
 
     /**
      * Gets a list of the shapes that can be used to launch a new Autonomous Exadata Infrastructure DB system. The shape determines resources to allocate to the DB system (CPU cores, memory and storage).