+++ bmc-database/src/main/java/com/oracle/bmc/database/DatabaseClient.java
@@ -2701,6 +2701,35 @@ public class DatabaseClient implements Database {
     }
 
     @Override
+    public ListAutonomousDbVersionsResponse listAutonomousDbVersions(
+            ListAutonomousDbVersionsRequest request) {
+        LOG.trace("Called listAutonomousDbVersions");
+        final ListAutonomousDbVersionsRequest interceptedRequest =
+                ListAutonomousDbVersionsConverter.interceptRequest(request);
+        com.oracle.bmc.http.internal.WrappedInvocationBuilder ib =
+                ListAutonomousDbVersionsConverter.fromRequest(client, interceptedRequest);
+        com.google.common.base.Function<javax.ws.rs.core.Response, ListAutonomousDbVersionsResponse>
+                transformer = ListAutonomousDbVersionsConverter.fromResponse();
+
+        final com.oracle.bmc.retrier.BmcGenericRetrier retrier =
+                com.oracle.bmc.retrier.Retriers.createPreferredRetrier(
+                        interceptedRequest.getRetryConfiguration(), retryConfiguration);
+        return retrier.execute(
+                interceptedRequest,
+                retryRequest -> {
+                    final com.oracle.bmc.retrier.TokenRefreshRetrier tokenRefreshRetrier =
+                            new com.oracle.bmc.retrier.TokenRefreshRetrier(
+                                    authenticationDetailsProvider);
+                    return tokenRefreshRetrier.execute(
+                            retryRequest,
+                            retriedRequest -> {
+                                javax.ws.rs.core.Response response = client.get(ib, retriedRequest);
+                                return transformer.apply(response);
+                            });
+                });
+    }
+
+    @Override
     public ListAutonomousExadataInfrastructureShapesResponse
             listAutonomousExadataInfrastructureShapes(
                     ListAutonomousExadataInfrastructureShapesRequest request) {