+++ bmc-loadbalancer/src/main/java/com/oracle/bmc/loadbalancer/model/ConnectionConfiguration.java
@@ -34,18 +34,30 @@ public class ConnectionConfiguration {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("backendTcpProxyProtocolVersion")
+        private Integer backendTcpProxyProtocolVersion;
+
+        public Builder backendTcpProxyProtocolVersion(Integer backendTcpProxyProtocolVersion) {
+            this.backendTcpProxyProtocolVersion = backendTcpProxyProtocolVersion;
+            this.__explicitlySet__.add("backendTcpProxyProtocolVersion");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonIgnore
         private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
 
         public ConnectionConfiguration build() {
-            ConnectionConfiguration __instance__ = new ConnectionConfiguration(idleTimeout);
+            ConnectionConfiguration __instance__ =
+                    new ConnectionConfiguration(idleTimeout, backendTcpProxyProtocolVersion);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
             return __instance__;
         }
 
         @com.fasterxml.jackson.annotation.JsonIgnore
         public Builder copy(ConnectionConfiguration o) {
-            Builder copiedBuilder = idleTimeout(o.getIdleTimeout());
+            Builder copiedBuilder =
+                    idleTimeout(o.getIdleTimeout())
+                            .backendTcpProxyProtocolVersion(o.getBackendTcpProxyProtocolVersion());
 
             copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
             return copiedBuilder;
@@ -71,6 +83,15 @@ public class ConnectionConfiguration {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("idleTimeout")
     Long idleTimeout;
+
+    /**
+     * The backend TCP Proxy Protocol version.
+     * <p>
+     * Example: `1`
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("backendTcpProxyProtocolVersion")
+    Integer backendTcpProxyProtocolVersion;
 
     @com.fasterxml.jackson.annotation.JsonIgnore
     private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();