+++ bmc-ons/src/main/java/com/oracle/bmc/ons/NotificationDataPlaneClient.java
@@ -18,7 +18,7 @@ public class NotificationDataPlaneClient implements NotificationDataPlane {
             com.oracle.bmc.Services.serviceBuilder()
                     .serviceName("NOTIFICATIONDATAPLANE")
                     .serviceEndpointPrefix("notification")
-                    .serviceEndpointTemplate("https://notification.{region}.oraclecloud.com")
+                    .serviceEndpointTemplate("https://notification.{region}.{secondLevelDomain}")
                     .build();
     // attempt twice if it's instance principals, immediately failures will try to refresh the token
     private static final int MAX_IMMEDIATE_RETRIES_IF_USING_INSTANCE_PRINCIPALS = 2;